# api/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "Problem": {
        "properties": {
          "detail": {
            "type": "string"
          },
          "id": {
            "description": "Identifier of the problem that occurred, stable across releases",
            "enum": [
              "OrganizationNameMissing",
              "UserAlreadyExist",
              "RegistrationDisabled",
              "PasswordTooShort",
              "OrganizationIsNotEmpty",
              "InvitedUserDoesNotExist",
              "InvitedUserAlreadyInOrganization",
              "ApplicationNameMissing",
              "InvalidRole",
              "EventTypeAlreadyExist",
              "EventTypeDoesNotExist",
              "UnauthorizedWorkers",
              "EventAlreadyIngested",
              "EventInvalidPayloadContentType",
              "EventInvalidBase64Payload",
              "EventInvalidJsonPayload",
              "LabelsAmbiguity",
              "InvalidDateRange",
              "AuthNoAuthorizationHeader",
              "AuthInvalidAuthorizationHeader",
              "AuthApplicationSecretLookupError",
              "AuthInvalidApplicationSecret",
              "AuthBiscuitLookupError",
              "AuthInvalidBiscuit",
              "AuthFailedLogin",
              "AuthEmailNotVerified",
              "AuthEmailAlreadyVerified",
              "AuthFailedRefresh",
              "AuthEmailExpired",
              "TooManyMembersPerOrganization",
              "TooManyApplicationsPerOrganization",
              "TooManyEventsToday",
              "TooManySubscriptionsPerApplication",
              "TooManyEventTypesPerApplication",
              "JsonPayload",
              "Validation",
              "NotFound",
              "InternalServerError",
              "Forbidden",
              "ServiceUnavailable"
            ],
            "type": "string"
          },
          "status": {
            "format": "int32",
            "type": "integer"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "detail",
          "id",
          "status",
          "title"
        ],
        "type": "object"
      },
...
      "RequestAttempt": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "delay_until": {
            "format": "date-time",
            "type": "string"
          },
          "event": {
            "properties": {
              "event_id": {
                "format": "uuid",
                "type": "string"
              },
              "event_type_name": {
                "type": "string"
              }
            },
            "required": [
              "event_id",
              "event_type_name"
            ],
            "type": "object"
          },
          "event_id": {
            "format": "uuid",
            "type": "string"
          },
          "failed_at": {
            "format": "date-time",
            "type": "string"
          },
          "http_response_status": {
            "format": "int32",
            "type": "integer"
          },
          "picked_at": {
            "format": "date-time",
            "type": "string"
          },
          "request_attempt_id": {
            "format": "uuid",
            "type": "string"
          },
          "response_id": {
            "format": "uuid",
            "type": "string"
          },
          "retry_count": {
            "format": "int32",
            "type": "integer"
          },
          "status": {
            "description": "Status of a request attempt. The 'type' field indicates the status variant. - waiting: {type, since, until} - Scheduled for future delivery - pending: {type, since} - Ready to be processed - in_progress: {type, since} - Currently being delivered - successful: {type, at, full_processing_ms} - Delivered successfully - failed: {type, at, full_processing_ms} - Delivery failed",
            "properties": {
              "at": {
                "description": "Timestamp when completed (present in successful, failed)",
                "format": "date-time",
                "type": "string"
              },
              "full_processing_ms": {
                "description": "Total processing time in milliseconds (present in successful, failed)",
                "format": "int64",
                "type": "integer"
              },
              "since": {
                "description": "Timestamp when the status started (present in waiting, pending, in_progress)",
                "format": "date-time",
                "type": "string"
              },
              "type": {
                "description": "Status type discriminator. One of: waiting, pending, in_progress, successful, failed",
                "enum": [
                  "waiting",
                  "pending",
                  "in_progress",
                  "successful",
                  "failed"
                ],
                "type": "string"
              },
              "until": {
                "description": "Timestamp until which waiting (only present in waiting status)",
                "format": "date-time",
                "type": "string"
              }
            },
            "required": [
              "type"
            ],
            "type": "object"
          },
          "subscription": {
            "properties": {
              "description": {
                "type": "string"
              },
              "subscription_id": {
                "format": "uuid",
                "type": "string"
              }
            },
            "required": [
              "subscription_id"
            ],
            "type": "object"
          },
          "succeeded_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "created_at",
          "event",
          "event_id",
          "request_attempt_id",
          "retry_count",
          "status",
          "subscription"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "biscuit": {
        "description": "Authentication using a Biscuit token (use the format `Bearer TOKEN`)",
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Core REST API of Hook0 — Open-Source Webhooks as a service for SaaS",
    "title": "Hook0 API",
    "version": "1.0.2"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/request_attempts/": {
      "get": {
        "description": "Retrieves webhook delivery attempts for an application. Each attempt shows the delivery status (pending, in_progress, successful, failed, waiting), retry count, and timestamps. Filter by event_id, subscription_id, date range, or event types. Paginated via Link header.",
        "operationId": "requestAttempts.list",
        "parameters": [
          {
            "in": "query",
            "name": "application_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "style": "form"
          },
          {
            "description": "Comma-separated event types",
            "in": "query",
            "name": "event.event_type_names",
            "schema": {
              "type": "string"
            },
            "style": "form"
          },
          {
            "in": "query",
            "name": "event_id",
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "style": "form"
          },
          {
            "in": "query",
            "name": "max_created_at",
            "schema": {
              "format": "date-time",
              "type": "string"
            },
            "style": "form"
          },
          {
            "in": "query",
            "name": "min_created_at",
            "schema": {
              "format": "date-time",
              "type": "string"
            },
            "style": "form"
          },
          {
            "in": "query",
            "name": "pagination_cursor",
            "schema": {
              "type": "string"
            },
            "style": "form"
          },
          {
            "in": "query",
            "name": "subscription_id",
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "style": "form"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RequestAttempt"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "biscuit": []
          }
        ],
        "summary": "List request attempts",
        "tags": [
          "Subscriptions Management",
          "mcp",
          "public"
        ]
      }
    },
    "/api/v1/request_attempts/{request_attempt_id}": {
      "get": {
        "description": "Retrieves a single webhook delivery attempt by ID, including delivery status, retry count, timestamps, and HTTP response status.",
        "operationId": "requestAttempts.get",
        "parameters": [
          {
            "in": "query",
            "name": "application_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "style": "form"
          },
          {
            "in": "path",
            "name": "request_attempt_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "style": "simple"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestAttempt"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "biscuit": []
          }
        ],
        "summary": "Get a request attempt by its ID",
        "tags": [
          "Subscriptions Management",
          "mcp",
          "public"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "https://app.hook0.com"
    }
  ]
}